auth/requirements.txt
asgiref==3.5.2
black==22.6.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==37.0.4
defusedxml==0.7.1
Django==4.1.3
django-allauth==0.51.0
django-debug-toolbar==3.6.0
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.13.1
drf-nested-routers==0.93.4
gunicorn==20.1.0
idna==3.3
Markdown==3.4.1
mypy-extensions==0.4.3
mysqlclient==2.1.1
oauthlib==3.2.0
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.12
whitenoise==6.2.0
django-anymail[sendinblue]
boto3==1.24.66
django-storages==1.13.1
dj-rest-auth==2.2.5
djangorestframework-simplejwt==5.2.1
django-extensions==3.2.0